Move `getMiddlewareRegex` to `route-regex`
