+++ packages/next/shared/lib/router/utils/index.ts
@@ -1,4 +1,3 @@
-export { getMiddlewareRegex } from './get-middleware-regex'
 export { getRouteMatcher } from './route-matcher'
 export { getRouteRegex } from './route-regex'
 export { getSortedRoutes } from './sorted-routes'
