+++ packages/next/shared/lib/router/utils/route-regex.ts
@@ -6,18 +6,6 @@ interface Group {
   optional: boolean
 }
 
-function parseParameter(param: string) {
-  const optional = param.startsWith('[') && param.endsWith(']')
-  if (optional) {
-    param = param.slice(1, -1)
-  }
-  const repeat = param.startsWith('...')
-  if (repeat) {
-    param = param.slice(3)
-  }
-  return { key: param, repeat, optional }
-}
-
 export function getParametrizedRoute(route: string) {
   const segments = (route.replace(/\/$/, '') || '/').slice(1).split('/')
 
@@ -129,4 +117,63 @@ export function getRouteRegex(normalizedRoute: string): RouteRegex {
     re: new RegExp(`^${result.parameterizedRoute}(?:/)?$`),
     groups: result.groups,
   }
+}
+
+export function getMiddlewareRegex(
+  normalizedRoute: string,
+  catchAll: boolean = true
+): RouteRegex {
+  const result = getParametrizedRoute(normalizedRoute)
+
+  let catchAllRegex = catchAll ? '(?!_next).*' : ''
+  let catchAllGroupedRegex = catchAll ? '(?:(/.*)?)' : ''
+
+  if ('routeKeys' in result) {
+    if (result.parameterizedRoute === '/') {
+      return {
+        groups: {},
+        namedRegex: `^/${catchAllRegex}$`,
+        re: new RegExp(`^/${catchAllRegex}$`),
+        routeKeys: {},
+      }
+    }
+
+    return {
+      groups: result.groups,
+      namedRegex: `^${result.namedParameterizedRoute}${catchAllGroupedRegex}$`,
+      re: new RegExp(`^${result.parameterizedRoute}${catchAllGroupedRegex}$`),
+      routeKeys: result.routeKeys,
+    }
+  }
+
+  if (result.parameterizedRoute === '/') {
+    return {
+      groups: {},
+      re: new RegExp(`^/${catchAllRegex}$`),
+    }
+  }
+
+  return {
+    groups: {},
+    re: new RegExp(`^${result.parameterizedRoute}${catchAllGroupedRegex}$`),
+  }
+}
+
+/**
+ * Parses a given parameter from a route to a data structure that can be used
+ * to generate the parametrized route. Examples:
+ *   - `[...slug]` -> `{ name: 'slug', repeat: true, optional: true }`
+ *   - `[foo]` -> `{ name: 'foo', repeat: false, optional: true }`
+ *   - `bar` -> `{ name: 'bar', repeat: false, optional: false }`
+ */
+function parseParameter(param: string) {
+  const optional = param.startsWith('[') && param.endsWith(']')
+  if (optional) {
+    param = param.slice(1, -1)
+  }
+  const repeat = param.startsWith('...')
+  if (repeat) {
+    param = param.slice(3)
+  }
+  return { key: param, repeat, optional }
 }
